FIX : pembetulan dari sheet

+++ app/Views/amalan_yaumi/index.php
@@ -54,19 +54,23 @@ function youtube($url){
                                                 <td><?= $row->judul_amalan_yaumi;?></td>
                                                 <td><?= $row->konten_amalan_yaumi;?></td>
                                                 <td align="center">
-                                                    <?php if ($row->status_amalan_yaumi == 1) : ?>
-                                                        <span class="badge bg-success text-light">Show</span>
-                                                        <hr>
-                                                        <a class="edit_status" class="btn btn-outline-secondary" href="/amalan_yaumi/edit_status/<?= $row->id_amalan_yaumi; ?>/0" onClick='return confirm("Yakin akan hide data amalan yaumi ini untuk user?")'><button type="button" class="btn btn-outline-secondary">Hide</button></a>
-                                                    <?php elseif ($row->status_amalan_yaumi == 0) : ?>
-                                                        <span class="badge bg-danger text-light">Hide</span>
-                                                        <hr>
-                                                        <a class="edit_status" class="btn btn-outline-secondary" href="/amalan_yaumi/edit_status/<?= $row->id_amalan_yaumi; ?>/1"><button type="button" class="btn btn-outline-secondary" onClick='return confirm("Yakin akan show data amalan yaumi ini untuk user?")'>Show</button></a>
+                                                    <?php if ($row->id_amalan_yaumi != 1): ?> <!-- untuk default amalan yaumi -->
+                                                        <?php if ($row->status_amalan_yaumi == 1) : ?>
+                                                            <span class="badge bg-success text-light">Show</span>
+                                                            <hr>
+                                                            <a class="edit_status" class="btn btn-outline-secondary" href="/amalan_yaumi/edit_status/<?= $row->id_amalan_yaumi; ?>/0" onClick='return confirm("Yakin akan hide data amalan yaumi ini untuk user?")'><button type="button" class="btn btn-outline-secondary">Hide</button></a>
+                                                        <?php elseif ($row->status_amalan_yaumi == 0) : ?>
+                                                            <span class="badge bg-danger text-light">Hide</span>
+                                                            <hr>
+                                                            <a class="edit_status" class="btn btn-outline-secondary" href="/amalan_yaumi/edit_status/<?= $row->id_amalan_yaumi; ?>/1"><button type="button" class="btn btn-outline-secondary" onClick='return confirm("Yakin akan show data amalan yaumi ini untuk user?")'>Show</button></a>
+                                                        <?php endif ?>
                                                     <?php endif ?>
                                                 </td>
                                                 <td width="150px" align="center">
-                                                    <a class="edit" class="btn btn-warning" href="/amalan_yaumi/edit/<?= $row->id_amalan_yaumi;?>" title="Edit Data"><button type="button" class="btn btn-warning"><i class="fas fa-fw fa-pen"></i></button></a>
-                                                    <a class="hapus" class="btn btn-danger" href="/amalan_yaumi/delete/<?= $row->id_amalan_yaumi;?>" onClick='return confirm("Yakin akan menghapus data amalan yaumi ini?")' title="Hapus Data"><button type="button" class="btn btn-danger"><i class="fas fa-fw fa-trash"></i></button></a> 
+                                                    <?php if ($row->id_amalan_yaumi != 1): ?> <!-- untuk default amalan yaumi -->
+                                                        <a class="edit" class="btn btn-warning" href="/amalan_yaumi/edit/<?= $row->id_amalan_yaumi;?>" title="Edit Data"><button type="button" class="btn btn-warning"><i class="fas fa-fw fa-pen"></i></button></a>
+                                                        <a class="hapus" class="btn btn-danger" href="/amalan_yaumi/delete/<?= $row->id_amalan_yaumi;?>" onClick='return confirm("Yakin akan menghapus data amalan yaumi ini?")' title="Hapus Data"><button type="button" class="btn btn-danger"><i class="fas fa-fw fa-trash"></i></button></a> 
+                                                    <?php endif ?>
                                                 </td>
                                             </tr>
                                         <?php endforeach; ?>
